npm install fix for NPM v3
NPM v3 changed how packages are installed into node_modules in a way that is
incompatible with the postinstall script we used to setup symlinks to the
top-level lib/ and test/ folders. Rather than keep this script up-to-date, let's
remove it and return to the less-than-pleasant relative `require` paths.

In the future, we can reorganize the lib/ and test/ folders to reduce the
amount of nesting.