Stop setting identity/Account SID in headers, URLs, etc.
SIP.js has a bug in their SIP URI encode/decode logic. Rather than work around
this, let's just use a UUID for the SIP From header and rely on Room Signaling
Protocol (RSP) for communicating identity. We don't check the SIP From header
anyway. We can also stop signaling the Account SID in our SIP headers, and we
don't need to use it when connecting to the WebSocket server.