Run integration tests in Chrome & Firefox w/ Karma
This commit runs our existing integration tests in Chrome and Firefox using
Karma. Some of the integration tests were tweaked to selectively override the
WebRTC implementation, PeerConnectionManager, or other APIs in order to work in
both Node and the browser. We also needed to update our SIP.js dependency to
workaround a SIP.js error when the CancelablePromise returned by `connect` is
`cancel`ed in the same tick. (JSDK-806, JSDK-807)