Stop using moz-prefixed WebRTC APIs
Rather than use the moz-prefixed WebRTC APIs in Firefox, we'll now use the non-
prefixed APIs. This should suppress the warning we have been receiving that

> WebRTC interfaces with the “moz” prefix (mozRTCPeerConnection,
> mozRTCSessionDescription, mozRTCIceCandidate) have been deprecated.

Note that we still do feature detection on the moz-prefixed APIs to determine
whether or not to load our Firefox RTCPeerConnection adapter. We can continue
to rely on the presence of these APIs until either of the following changes
land in Firefox:

* [Remove legacy support for moz-prefixed WebRTC interfaces eventually](https://bugzilla.mozilla.org/show_bug.cgi?id=1241283)
* [[meta] Remove all deprecated interfaces in WebRTC and MediaCapture and Streams eventually.](https://bugzilla.mozilla.org/show_bug.cgi?id=1241290)