Prefer getSenders/getReceivers in Firefox
In Firefox, we should prefer the `getSenders` and `getReceivers` APIs to the
`getLocalStreams` and `getRemoteStreams` APIs in order to suppress the following
warning:

> RTCPeerConnection.getLocalStreams/getRemoteStreams are deprecated. Use
> RTCPeerConnection.getSenders/getReceivers instead.

However, until Chrome adds support for these APIs, it makes sense for our
RTCPeerConnection adapter to continue providing support for the deprecated APIs.
So we continue to provide these APIs in lib/webrtc/rtcpeerconnection/firefox.js,
although the APIs behave a little differently: we don't return the actual
MediaStream objects negotiated. Instead, we construct a new one at each call to
`getLocalStreams` or `getRemoteStreams` and stuff each MediaStreamTrack into
that new MediaStream.

This is sufficient for our use of the WebRTC APIs, although it wouldn't be
a good fit for others.